+++ resolvers/mutation-resolver/mutations/validate-functions.js
